requirements.txt
absl-py==0.11.0
cachetools==4.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
cmake==3.18.2.post1
configparser==5.0.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
docker-pycreds==0.4.0
easydict==1.9
future==0.18.2
fvcore==0.1.2.post20210115
gitdb==4.0.5
GitPython==3.1.9
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.1
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
iopath==0.1.2
joblib==0.17.0
kiwisolver==1.2.0
kornia==0.4.0
Markdown==3.3.3
matplotlib==3.3.2
MulticoreTSNE==0.1
networkx==2.5
numpy==1.19.2
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
pandas==1.1.2
pathtools==0.1.2
Pillow==7.2.0
plotly==4.11.0
portalocker==2.0.0
promise==2.3
protobuf==3.13.0
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
sentry-sdk==0.18.0
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
subprocess32==3.5.4
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.0
typing-extensions==3.7.4.3
urllib3==1.25.10
wandb==0.10.4
watchdog==0.10.3
Werkzeug==1.0.1
yacs==0.1.8
zipp==3.4.0
